requirements.txt
azure-functions
pymupdf==1.24.9
